Add GaMD conf files
Gaussian accelerated MD is pretty cool.
Note, you will have to include the EQ-e- stage in your yaml file, and change the production protocol name to GaMD (and update the template) in the yaml file.
If you have a particularly large system, edit the EQ-e- .conf file in templates, and update to increase the number of equilibration steps (READ THE AMBER MANUAL!).